requirements.txt
selenium
flask
pillow
werkzeug